requirements.txt
openai-whisper
tqdm
google-generativeai
python-dotenv
requests
